fix(eval): pinned-regressions lane migrates the runtime self-DB before its ORM checks (#2190)
The eval-pinned-regressions pre-push lane red on this branch with
`OperationalError: no such column last_used_at`. Root cause: the
corpus's ORM-backed checks query the runtime-resolved self-DB, and a
worktree's auto-isolated DB is seeded from a snapshot of the canonical
DB at its (possibly stale) schema with migrations never applied — so a
PR adding a migration leaves that DB missing the new column/table and
every ORM check raises OperationalError.

Add an in-process schema pre-flight that brings the runtime self-DB
current before any ORM check, reusing the sanctioned, idempotent,
non-destructive `schema_guard.migrate_self_db` against the exact
connection the checks read. A migrate failure fails the corpus LOUD
(fail-closed) — never a silent pass against a half-migrated DB. This is
the structural fix: migrating the operator control DB instead is refused
from a worktree by design (CanonicalDBFromWorktreeError), and would not
have covered the per-worktree isolated copy that actually fails.

Split the corpus into regression_corpus_schema (the pre-flight) and
regression_corpus_predicates (the _check_* bodies) to keep each module
under the module-health LOC cap; the migration-fork predicate stays in
regression_corpus so its anti-vacuous test can patch _count_core_leaves
on that namespace.
